Add known dependents file
The idea is that if we find one of these packages as a direct dependency, we can infer govuk-frontend's version.

This would rely on a function which checks which package manager we're dealing with in the first instance, and therefore which files to check for dependencies.

We could include manual ports here, and potentially run separate dependents data for them to get an idea of who's using the ports that don't use our package underlying and how much of a market there is.